refactor: moved column utils to shared file
